Move Badge to com.trafi.ui.component package.
