+++ android/routes/ui/src/main/java/com/trafi/routes/ui/RouteSegment.kt
@@ -1,7 +1,8 @@
 package com.trafi.routes.ui
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.defaultMinSizeConstraints
 import androidx.compose.material.AmbientContentColor
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -19,9 +20,9 @@ import com.trafi.core.model.*
 import com.trafi.routes.ui.internal.endTimeMillis
 import com.trafi.routes.ui.internal.startTimeMillis
 import com.trafi.routes.ui.mock.RouteSegmentPreviewParameterProvider
-import com.trafi.ui.Badge
-import com.trafi.ui.BadgeInfo
-import com.trafi.ui.BadgeType
+import com.trafi.ui.component.Badge
+import com.trafi.ui.component.BadgeInfo
+import com.trafi.ui.component.BadgeType
 import com.trafi.ui.theme.MaasTheme
 import com.trafi.ui.theme.internal.toColor
